requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
docopt==0.6.2
Flask==2.0.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
pycparser==2.21
pycryptodome==3.12.0
pyOpenSSL==21.0.0
python-dotenv==0.19.2
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.29
urllib3==1.26.8
Werkzeug==2.0.2
yarg==0.1.9
zipp==3.7.0
